app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
PyJWT==2.7.0
gunicorn==20.1.0
python-jose>=3.0.0
python-multipart

pyparsing==3.1.1
python-dateutil~=2.8.2
aiohttp~=3.8.5
betterlogging~=0.2.1
requests==2.28.1